Add a dummy stat implementation for OpenSSL
The OpenSSL build requires a stat implementation, the file
crypto/store/loader_file.c calls it. On the other hand, we don't need
the loader_file for the SVSM, so this patch adds a dummy stat
implementation.

The bits/alltypes.h, bits/stat.h and sys/stat.h are imported from musl.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>